+++ doc.go
@@ -1,4 +1,4 @@
-// Package slogx provides extensions to the standard slog package.
+// Package slogx is an enhanced version of the built-in log/slog package.
 //
 // TODO: add full documentation across all packages with examples...
 package slogx