+++ handler/json.go
@@ -19,9 +19,7 @@ type jsonHandlerOptionsContext struct{}
 // JSONHandlerOptions holds the options for the JSON handler.
 type JSONHandlerOptions struct {
 	// Level is the minimum log level to write to the handler.
-	//
-	// By default, the level will be set to slog.LevelInfo if not supplied.
-	Level slog.Leveler
+	Level slogx.Level
 
 	// RecordFormatter specifies the formatter to use to format the record before writing it to the writer.
 	//
@@ -37,7 +35,7 @@ type JSONHandlerOptions struct {
 // DefaultJSONHandlerOptions returns a default set of options for the handler.
 func DefaultJSONHandlerOptions() JSONHandlerOptions {
 	return JSONHandlerOptions{
-		Level:           slog.LevelInfo,
+		Level:           slogx.LevelInfo,
 		RecordFormatter: formatter.DefaultJSONFormatter(),
 		Writer:          os.Stdout,
 	}
@@ -74,9 +72,6 @@ type jsonHandler struct {
 // NewJSONHandler creates a new handler object.
 func NewJSONHandler(opts JSONHandlerOptions) *jsonHandler {
 	// set default options
-	if opts.Level == nil {
-		opts.Level = slog.LevelInfo
-	}
 	if opts.Writer == nil {
 		opts.Writer = os.Stdout
 	}
@@ -124,6 +119,11 @@ func (h *jsonHandler) Handle(ctx context.Context, r slog.Record) error {
 	return err
 }
 
+// Level returns the current logging level that is in use by the handler.
+func (h jsonHandler) Level() slogx.Level {
+	return h.options.Level
+}
+
 // Shutdown is responsible for cleaning up resources used by the handler.
 func (h jsonHandler) Shutdown(continueOnError bool) error {
 	if w, ok := h.options.Writer.(io.WriteCloser); ok {
@@ -164,4 +164,17 @@ func (h *jsonHandler) WithGroup(name string) slog.Handler {
 		newHandler.activeGroup = name
 	}
 	return newHandler
+}
+
+// WithLevel returns a new handler with the given logging level set.
+func (h jsonHandler) WithLevel(level slogx.Level) slogx.DynamicLevelHandler {
+	options := h.options
+	options.Level = level
+	return &jsonHandler{
+		activeGroup: h.activeGroup,
+		attrs:       h.attrs,
+		groups:      h.groups,
+		options:     options,
+		writeLock:   h.writeLock,
+	}
 }